Issue 5971: programmatic triggering of NetExport
An export can be triggered with
`window.NetExport.triggerExport("token")()`. Automatic export has to
be enabled for this function to be available. However, it is possible
to disable unattended automatic export by setting the timeout to 0.

Since we don't want any page to trigger an export, `token` is a
security token that should be kept secret and is set by the user in
`extensions.firebug.netexport.secretToken`. If the token is incorrect,
it is not possible to call the final function.